{TRANSFER] PyCharm

+++ _system_.fld/lib/miranda_log/__init__.py
@@ -19,11 +19,14 @@ def log(action, description):
         file.write(msg)
 
 def clear(d=False):
+    idPath = os.path.join(miranda_folder.root(), "_system_.fld", "var", "ID.var")
+    id = int(open(idPath, "r").read())
     path = os.path.join(miranda_folder.root(), "_logs_.fld")
     pattern = re.compile(r"_log_ID\d+_.log$")
     for filename in os.listdir(path):
-        if pattern.match(filename):
+        if pattern.match(filename) and filename != f"_log_ID{id}_.log":
             fPath = os.path.join(path, filename)
             os.remove(fPath)
+            log("INFO", "DELETED FILE: %s" % fPath)
             if d:
                 print(":: REMOVED FILE: %s" % fPath)
